edit RabbitMQ license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit modifies a few of RabbitMQ's docs so that Licensee is able
to recognize the repository's license type. It removes LICENSE so that
LICENSE-MPL-RabbitMQ is the sole license document in the repo. The
contact info that was stored in LICENSE has now been moved to a new
"Licensing" section in the README. Also, LICENSE-MPL-RabbitMQ has been
updated to store the MPL 2.0 license (instead of the MPL 1.1 license).

Collectively, these changes allow Licensee to successfully identify the
license type of RabbitMQ as MPL 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>