Be more defensive about what output rabbitmqctl eval produces here
OTP 21 has a new logging subsystem and extra logging added to the kernel app
to accompany the asynchronous automatic node connection feature. This breaks
this eval command that asks the node for some of its effective paths (e.g.
the plugin dir).

Since we know the set of lines we are interested in, we can filter out
everything else.

Pair: @dumbbell.

Part of #1616.

[#157964874]