Don't await for error_logger handlers to initialise
It will never be started on OTP 21.

The problem this piece of code originally was added to solve
is not very relevant now that we use Lager. Arguably connection
rate throttling should be enforced elsewhere, be it Ranch or a proxy/load balancer.

In any case, this particular implementation worked around an error_logger
issue that is no longer problematic for us, and eventually won't be problematic
for the entire ecosystem because OTP 21 introduced a completely new logger.

Cherry-picked per discussion with @dumbbell and @lukebakken.

[#157964874]

(cherry picked from commit a54b44fd8d24af90d8df301658fba4775f957bc8)