rabbirmq-run.mk: Use rabbitmqctl(1) instead of erl_call(1)
erl_call(1) is broken in Erlang 21.0-rc.2 when we use it to evaluate
some code on a remote node. The issue was reported upstream but we can
replace erl_call(1) by `rabbitmqctl eval` to achieve the same result.

The output is still filtered using sed(1) to ensure it remains
unchanged, in case testsuites expect a particular format.

We still continue to use erl_call(1) to stop a node because we can't
achieve the same behavior using rabbitmqctl(1). This use of erl_call(1)
is working with Erlang 21.0-rc.2.

[#157964874]

(cherry picked from commit 0312c6c5ed64f6aca3f30f968f4d4c08fdc2002b)