Use %2F instead of %2f to percent-encode the `/` vhost
The meaning is the same. However, the RFC recommends to use uppercase in
percent-encoded characters. Morever, the new URI normalization in Erlang
21 would transform the percent-encoded characters to uppercase.

And because our `rabbit_web_dispatch_util:relativise()` uses exact
matches, it will consider e.g `%2f` and `%2F` to be different
characters. Therefore when we use Erlang 21, the relative redirection
returned by the management plugin will be different than the one
returned with Erlang 20 if we use lowercase percent-encoded `/`.

Thus the easy fix is to use uppercase.

A better fix would be to improve `relativise()` and the testsuite to
compare percent-decoded components.

[#157964874]