rabbit_mgmt_http_SUITE: Percent-encode `^` in URLs
This character is neither reserved nor unreserved and thus is an unsafe
character according to RFC 1738 and RFC 3986. Therefore it should be
percent-encoded.

Many implementations accept and/or produce URLs with raw `^`. However,
the URI parser in Erlang 21 is strict and will refuse that character.
Therefore, we now percent-encode it in the testsuite.

[#157964874]

(cherry picked from commit daf3811cdcffd5cddc0b2ccb455830b7b1eb9dcb)