Make rabbit_event:notify NOT fail if rabbit_event is not started yet.
If resource alarm is triggered during the boot process it will send
an event via rabbit_event:notify. This can crash the node if rabbit_event
is not started yet. rabbit_event starts after rabbit_alarm and any alarms
on boot were crashing the node.

Calling gen_event:notify with {rabbit_event, node()} is the same as with
rabbit_event except it does not fail.