Avoid duplicate keys in rabbit_mgmt_format:strip_pids/1
For proplists that represent queues the key set is fixed,
so special casing 'node' should be good enough. For other entites
we take a heavier handed approach and remove all duplicates from
the input proplist. The overhead of this with 1K entities seems
to be low single digit % in an end-to-end test that involves an HTTP
client.

Once we switch that code to use maps, the workaround won't be
necessary at all.

Closes rabbitmq/rabbitmq-management#601.

[#159578855]

(cherry picked from commit 52837c253729da172469f1827cb2937160eee8c6)