Handle generic data structures in priority queue info callback
The priority queue implementation makes assumptions over the
data format returned by the underlying backing queue
`info` callback.

This causes priority queues to crash if used together with other
modules implementing the `rabbit_backing_queue` behaviour and
returning additional data with the `info` callback.

As we cannot predict the format of the data the backing queue
`info` callback could return, we just replace it with the undefined
keyword.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>