Update bucket numbers for each queue just once
There can be more than one bucket per queue, so we ended up with
potential extra updates that resulted in incorrect (negative) bucket numbers.

As part of this we considered using an alternative data model:
folding two tables into one that looks like this:

exchange => map(bucket => queue)

This would greatly simplify binding management and be roughly as
efficient for routing except for one thing: updating the rest of the ring
will be a linear operation over all buckets. So that change alone
would be insufficient but remains to be an idea to improve on
in the future (e.g. by using a tree).

Pair: @dcorbacho.

References #37, #38.

[#159822323]

(cherry picked from commit 5cab6eb426e683a776387ac5f141ba7c6687d919)