Add localhost as SAN to server certificate
This helps for Java client hostname verification tests on CI: the CI
containers resolve the hostname to an external IP address and the broker
doesn't accept the connection for guest because it's not from localhost.
By using localhost in the server certificate SAN, hostname verification
is enforced and the connection is from localhost.