Don't error when a queue that's being deleted has no bindings to remove
It can happen due to retries of (currently not guaranteed to be idempotent,
which is a separate issue in the works) binding removal. Since both
the queue and its bindings are undergoing removal, don't fail when
there's nothing left to be removed for the current transaction [attempt].

This avoids obscure and non-actionable errors in the log ({error, not_found}).
Note that the error is also not handled by the callers and ignoring them
is the only reasonable course of action that I can think of.

Per discussion with @dumbbell @dcorbacho @hairyhum.