Idempotent binding removal
Optimizations introduced in #1589 caused the removal of bindings to
be non-idempotent, as the removal of routing information with dirty
deletes did not allow for full transaction rollback. This commit
reverts that change, losing some of the performance improvements in
favour of data correctness.

[#160100569]

(cherry picked from commit 495ead38956b2d1bcde5db071118ca4b50ab2812)