Use absolute link for rabbitmqadmin
If a user is browsing the cli page via e.g. http://localhost:15672/cli,
the link will become http://localhost:15672/rabbitmqadmin since there is
no `/` after `/cli`. If a user is browsing the cli page via e.g.
`http://localhost:15672/cli/ the link will become
http://localhost:15672/cli/rabbitmqadmin and it will work as expected.

The easiest fix was to use an absolute link instead and not depend on
the final `/`.