Ensure that login success clears a warn popup
If you enter invalid credentials and attempt login, you will see the warning popup show. Then, do not close it but log in with valid credentials. The popup remains. You can close it via the "Close" button but this change will remove the popup for you

Fixes #609, reported in rabbitmq/rabbitmq-server#1704