Reply `reject_publish` when queue marks a message as duplicate
The `rabbit_backing_queue` behaviour defines a `is_duplicate` callback
to signal the message about to be published is a
duplicate. Nevertheless, the broker logic does not take any action if
so.

This becomes problematic if the client enables publisher confirmation
as neither a confirm nor a reject is provided as response if a message
gets dropped as duplicate.

This patch ensures the client receives a rejection if a queue marks
the delivered message as duplicate.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>