Implement exclusive consumer (WIP)
This commit introduces exclusive consumer. A queue is declared with the
x-exclusive-consumer argument set to true and it will dispatch messages
to only one consumer at a time. This exclusive consumer is typically
the first one to consume from the queue. Other consumers are simply
ignored. If the exclusive consumer is cancelled, another consumer is
picked up randomly.

Subsequent commits will handle other cases than simple cancellation
(e.g. the owning channel of the exclusive consumer is closed).

[#161090309]

References #1743