Quorum queue: fix crash bug after reconnection
When a consumer node is disconnected the quorum queue will return all
outstanding messages for the consumer to the queue as it cannot know if
the consumer will ever come back or not and cannot leave things checked
out forever. If the consumer node then reconnectes and the consumer
channel is still alive it may try to query for missing deliveries and
this previously crashed if the messages had been returned to the main
queue. This fix makes the query safer as well as handling missing
messages better.