Makefile: Make the source archive build reproducible
The easy part was to figure the latest Git commit date and use it as the
last modification date for all files in the archive. We also pass a few
flags to tar(1), in particular to set the owner and group to 0.

The hard part was again Elixir/Mix. Mix Hex component requires a cache
file in order to build things offline (no idea why). This cache file is
an ETS table. There two issues with this table:
  * It contains a timestamp of the last time the cache was updated.
  * `ets:tab2file()` produces a different file each time it's called,
    even though the entries are the same.

So in order to be able to reproduce this cache file, the content of the
ETS table is dumped as an Erlang term to a text file. Then, when
RabbitMQ is compiled, the first step is to recreate the ETS table from
that Erlang term.