Make mandatory based on route data only
Instead of waiting for a mandatory_received message from the queue the
mandatory result is calculated in the channel based on the routing
result only. This may seem like a weakening of the mandatory semantics but
considering that the mandatory_received message is returned _before_ the
message is enqueued and/or persisted in the queue it doesn't actually
open up any further failure scenarios.

[#163222515]