Use {active,once} instead of async_recv
A test had to be changed because closing the socket locally
does not trigger the condition we expect. It happened to work
before because async_recv is using a low level interface but
it wasn't really testing the right thing.

(cherry picked from commit d76750437bef553f2375f393aee5ca65b1bc86dd)