Change quorum queue delete timeout
The default timeout of net_ticktime + 5000 ms was much too long and made
the delete operation in the management UI timeout and not complete
properly. A simple 5000ms works better and this commit also handles the
case where the delete command was committed but the leader never shut
down due to a server being unavailable. This also falls back to forced
deletion.

[#163062078]