Quorum queue queue length limit by byte size and number of messages
Only drop-head strategy.

This necessitated the change of rabbit_fifo prefix messages from a tuple
of integers representing the number of returned vs enqueued messages
that have already been processes and thus don't need to include message
bodes in the snapshot to a tuple of lists of the sizes of each message.
This change will have some performance impact as the snaphots will now
be larger than before but as they still won't contain message bodies at
least the sizing is fixed. Decreased the frequency as snapshots points
are prepared so somewhat make up for this.

[#161247380]