RestartVhostCommandTest stability and efficiency improvements
Virtual host process tree needs a bit more time to start and stop/
update state on failure.

Before this test a slower 5 year old machine had a 20% failure test rate
with this test (2 out of 10 attempts), with this change it has a %0
failure rate (0 out of 50).

While at it, only set up the vhost for tests that will actually hit
the test node.