Add make target for displaying GitHub release stats
E.g.:

    ●  Release v3.7.10 [Latest]
         Assets: 36
         Downloads: 47.67k
         Date: 1/7/2019
         Author: @pivotal-rabbitmq-ci

There is an interactive variant, in case one wants to try out different flags.

Both targets depend on Docker, so that the host on which they run does
not get polluted with node.js & friends. All containers use a volume for
the node_modules so that subsequent runs are faster.

Those that run *BSD or Linux might want to adapt this for their
platform. This feels good enough as a first spike, which is yet to
survive the meticulous @dumbbell or the veteran @michaelklishin