version command test: remove an environment-specific test
`rabbit_misc:version/0` requires rabbit.app to be on code path
and loaded which is not always the case on the test node.

The command is so simple that this doesn't introduce enough risk
to justify spending more time on it.

(cherry picked from commit 99eaec4a8e927b8eb46c98e01ab80913f4a044ab)