Queue ownership check: provide more details in the messaage
See #1887 for context. When an exclusive queue is redeclared with
the exclusive property set to `false`, the code considers it to be
an ownership check. This is a leaked implementation detail that's
been around for years, so changing it might do more harm than good.

What we can do is provide a bit more information about when the
check might fail in the message.