Make sure that publishes to only dead or unaccessible queues return nack.
Channel is counting unacked messages in a remove-only data structure `dtree`.
Each published message id is associated with a list of queues where it was routed to.
On confirm of queue failures queues were removed from the list
As soon as there are no queues in the list - the message can be confirmed.

This meant that if all queues fail with "not abnormal" reasons - the message may be
confirmed, but not enqueued.

This patch adds confirmation tracking for publishes, so if a message was confirmed
by at least one queue it was routed to, it can be confirmed as soon as all queues
confirm or fail. If all the queues fail - it will be rejected.

[#163952410]