rabbitmq-env: Pass remote node name ro child rabbitmqctl(8)
We use rabbitmqctl(8) inside `rabbitmq-env` to query the settings of the
remote nodes (path to plugins, feature flags file, the database, etc).
However, before this patch, we didn't pass the name of the remote node
as specified by the user with `-n`. Therefore, the default node name was
used (`rabbit@$hostname`) and that node may not exist at all.

This caused the executed script to run with possibly incorrect settings.
In particular, this prevented rabbitmq-plugins(8) from working properly
on a node started with `gmake run-broker`.

This should now be fixed because we extract the remote node name from
the command line arguments and pass it to the child rabbitmqctl(8).

(cherry picked from commit d32660c5fc2abc6250ff9df2393eef2d407e7351)