Keep track of partially confirmed messages per queue.
The failure handling is designed to handle failures of mirrored queues.
If a queue mirror process is disconnected we don't want to reject the message.

There is no need to tolerate queue failures in other scenarios and if all
mirrors fail before confirming - that's also not a good delivery.

Make partial confitmations handling keep track of queues to distinguish
between mirrors failing and unmirrored queues failing.

[#163952410]