Fix rabbit_fifo poison handling
By ensuring the delivery count is retained when "dehydrating" the state
in preparation for snapshotting. Now the entire message header map is
stored which will take additional space w.r.t to keynamne duplication
although this can be optimised.

Also updated the property test to generate fake pids for multiple nodes
so that multi node scenarios are better covered.

[#163513253]