Less aggressive locking for bindings add/reomve.
All binding operations aquire special locks for destination and source.
This is done to make sure bindings cannot be added or removed while
being cleaned up.

This means that adding and removing bindings for the same records may
conflict.

To avoid conflict the locks aquired during add/remove are now read locks.

This should improve performance in case of concurrent bindings creation
for the same resource.