rabbit_ct_broker_helpers: Don't set secondary Umbrella env. if it's not configured
Before this patch, we were always setting the secondary
Umbrella-specific environment variables (like `DEPS_DIR` and paths to
`rabbitmq*` scripts) based on the number of the node, regardless of the
fact that a secondary Umbrella was configured or not.

This was fine when used inside an Umbrella becase the computed paths
were working by chance. However, when using a standalone clone of
rabbitmq-server, the computed paths to `rabbitmq*` paths were incorrect,
leading to the odd-numbered nodes to fail to start.

The fix is easy: do not set the secondary Umbrella-specific environment
variables when there is none configured.