rabbit_fifo_SUITE: Fix include of `rabbit_fifo.hrl`
We cannot use `-include_lib()` here because we are inside the
application.

It happened to work by chance because we use the Umbrella to work on all
projects, so rabbitmq-server is also inside `DEPS_DIR`. But when the
project is tested outside of the Umbrella, the testsuite failed to
compile.