rabbit: Compute `totals` in status() only if RabbitMQ is running
The `app_management` testcase checks that status() works even with
RabbitMQ stopped (but the node is running). To restore that fact,
`totals` are only computed if RabbitMQ is running. Otherwise the list of
`totals` is empty.

The problem reported by the testcase without this patch was that the
`rabbit_vhost` Mnesia table was unavailable:

    {exit,{aborted,{no_exists,rabbit_vhost}}}