close connections properly even if init step failed
Connections are not closed gracefully if destination or source
initialisation throws an exception.
Connections are created a bit earlier before src/dst are initialised.
If for any reason init phase fails, connection pids will be lost and
shovel protocols won’t be able to close them as they are not preserved
within a state.