rabbit_feature_flags: Export initialize_registry/1 for the testsuite
Hiding the export behind the `TEST` macro should work on the paper.
Unfortunately, when we compile a test plugin as part of the
`feature_flags_SUITE` testsuite, `rabbit` is recompiled without the
`TEST` macro. This causes the testsuite to fail with an `undef`
exception because initialize_registry/1 is not exported anymore.