Correctly recognise when access is denied for user
So far, the "deny" resopnse from an HTTP authentication server was parsed
as a string while an atom was expected. As a result, if the server denied
a user it simply happend to work as the plugin would return an error
because of the type mismatch and the user would not be accepted anyway.