New setting to confirm messages on ack.
Make it possible to configure queues to confirm messages either
on enqueue (default) or on ack.
This will make it possible to communicate to the publisher that the
message was processed by consumer.
For example in federation this will mean that the message is successfully
published on the downstream (1 level).

Configurations:
queue argument x-confirm-on
policy confirm-on

Values:
enqueue - old behaviour, confirm on enqueue (default)
ack - new behaviour, confirm when message is acked or dropped

TODO:
- There may still be some ways for a message to leave the queue
without acking. Need to find them all.

- Dead lettering and dropping of the message should probably not
confirm, but reject the message.

- When changing the configuration using policies to `enqueue`,
all messages will be confirmed. Need to implement some sort of
enqueued messages tracker to only confirm enqueued messages.

- Quorum queues