New setting to confirm messages on ack.
Make it possible to configure queues to confirm messages either
on enqueue (default) or on ack.
This will make it possible to communicate to the publisher that the
message was processed by consumer.
For example in federation this will mean that the message is successfully
published on the downstream (1 level).

When a message is dropped (ttl, =false or message limit) - it will be rejected.

Configurations:
queue argument x-confirm-on
policy confirm-on

Values:
enqueue - old behaviour, confirm on enqueue (default)
ack - new behaviour, confirm when message is acked or dropped