rabbitmq-mix.mk: Ensure `$HOME` does not come from the command line
... by using the `override` directive to enforce our internal value.
As a reminder, this a hack to convince mix(1) to work offline and use a
copy of the Hex.pm cache provided in the source archive.

Note that this file was moved from rabbitmq-common. It has no use there
and it avoids the need to go through the entire CI pipeline if we want
to change this file.

The `stats.mk` file was also moved to the new `mk` subdirectory to avoid
file pollution at the root of this project.

(cherry picked from commit 68213819aa4a6738cb1762b6597d211171f6ac9e)