Functions to support log streaming to a CLI node.
To stream log updates to remote node there will be a process started
on the server node to read the log file and send its contents.
Unfortunately it's hard to detect the file to be deleted and the streaming
process will try indefinitely to get more data from the deleted file.
Streaming process is linked to the CLI process and should eventually stop.