rabbitmqadmin: accept 102 Processing responses
Follow-up to #711, closes #715.

This avoids an exception when rabbitmqadmin receives a 102 Processing
response. However, this does not make the operation wait for
the import to complete, so we treat it as we would a 202 Accepted
response.

The sequence of responses for a long running import operation is
something like this (in pseudo code, not actual HTTP 1.1):

102 Processing
102 Processing
102 Processing
102 Processing
204 No Content

Unfortunately Python's httplib implicitly assumes that there can only
be one response per request, and its state machine throws exceptions
when the user attempts to consume multiple responses on a persistent
connection.