rabbit_mnesia: Be more flexible in is_virgin_node()
Before, if the directory was not empty, the function required that all
three ignored files were present in that directory. If only two were,
then the test would fail and the node would not be considered as
pristine.

Now, we take the list of files in that directory and remove the files to
ignored. If the result is empty, the node is new.

(cherry picked from commit 8600e543e384470f972d81238a94449433ea0645)