Makefile: Remove all useless bits
... which were copy-pasted from `rabbit_common` probably.

They were all ineffective because:
* this application was never published to Hex.pm so far
- the `$(BUILD_DEPS)` variable is overwritten in `development.pre.mk`,
  meaning that `rabbit_common` was never used as a dependency.

However in the Umbrella, Erlang.mk plugins were implicitly loaded
because they were available by chance. This went well so far, but broke
in the rabbitmq-server(8) replacement experiment because one of the
plugin started to rely on CLI scripts and escripts when running tests.
This is useless for this application anyway, so getting rid of the
plugin inclusions fix the problem.

(cherry picked from commit 2a58ccc73567908b77dec40f26434627c3d7a591)