rabbit_amqqueue: Return empty list in consumers/1 when the queue is gone
This seems to happen in our CI quite frequently with
`queue_parallel_SUITE`. In this testsuite, testcases are executed in
parallel. At least the `basic_cancel` testcase queries all consumers of
all queues: if an unrelated testcase finishes in parallel and thus
deletes its queue(s), the call to `consumers_all/1` crashes.