unit_inbroker_non_parallel_SUITE: Mark log files as writable before deleting them
On Unix, the system checks the permissions of the directory holding a
file to determine if a file can be removed or not: if there is write
permission on a directory, a file in this directory can be removed, even
though the file itself might be read-only.

On Windows however, a read-only file cannot be removed: trying to do so
returns `EACCES`.

Therefore, in the `log_management` testcase, after testing the behavior
with read-only log files, we must mark them as writable again before we
clean them up. Otherwise, the testcase fails on Windows.