Refer to actual service name in example pods
The example worked as it was but it was kinda misleading. We should not
be refering to SERVICE ACCOUNT name but to SERVICE name when setting
`K8S_SERVICE_NAME`, `K8S_HOSTNAME_PREFIX`, and `RABBITMQ_NODENAME`. The
example simply worked because the service account and the service happen
to have the same name. Change either the name of the service account OR
the name of the service in the original example and see it fail.