Debian: support Erlang 22.x
but not 23.0. By being overly conservative with minor version numbers
we repeatedly run into the same problem:

 * A RabbitMQ release comes out and marks 22.1 as unsupported
 * Erlang/OTP 22.1 comes out the next day
 * Debian users cannot install this combination which the docs unintentionally
   recommend by using the "latest 22" apt compoonent.

Minor versions of 21.x and 22.x have been compatible with the original
major release, so let's switch to a more optimistic version range
unless we have evidence of incompatibilities in recent minor releases.