Extract __inputs from Grafana dashboards
While __inputs are required for the dashboards to work in environments
where Prometheus is not the default datasource, it breaks the local
development flow. In other words,
https://github.com/rabbitmq/rabbitmq-prometheus/commit/9aa22e1895055a0c132941a4367e7b6a68bc4c22
prevents `make metrics overview` from working as designed.

We are going to add shortly a simple way of converting the local
dashboards into a format that can be imported in Grafana and will work
when Prometheus is not the default datasource (e.g. when using
https://github.com/coreos/kube-prometheus)

Long-term, these dashboards will be available via grafana.com, which is
the preferred way of consuming them.

cc @mkuratczyk