Add make targets for import-friendly Grafana dashboards
To get an import-friendly RabbitMQ Overview dashboard, run the following
command:

    make RabbitMQ-Overview.json

On macOS, to send this output to clipboard:

    make RabbitMQ-Overview.json | pbcopy

This is the preferred alternative to
https://github.com/rabbitmq/rabbitmq-prometheus/commit/9aa22e1895055a0c132941a4367e7b6a68bc4c22
See
https://github.com/rabbitmq/rabbitmq-prometheus/commit/dae49b5c080bc5572d737c40f293d96dd53f9b17
for more context. cc @mkuratczyk

This commit introduces a few other somewhat related changes:

* BASH autocompletion for make targets - make ac
* descriptions for all custom targets - make h
* continuous feedback loops for ac & h targets - make CFac

I would really like to see some of the above features be part of
erlang.mk. What do you think @essen? Anything in particular that you
would like me to PR?

@dumbbell, my other Make partner-in-crime, may be interested in
discussing the above ; )

== LINKS

* https://medium.com/@lavieenroux20/how-to-win-friends-influence-people-and-autocomplete-makefile-targets-e6cd228d856d
* https://github.com/Bash-it/bash-it/blob/master/completion/available/makefile.completion.bash