Support x-queue-type argument
To be able to use quorum queues. Note deliveries from a classic queue
come with a context to inform back the queue about the handling of the
message (credit flow). Deliveries from a quorum queue do have this
context because quorum queues do not use credit flow. So supporting
quorum queues is not just about propagating the type header to AMQP
but needs also to handle both kinds of deliveries.

Fixes #138