quorum_queue_SUITE: Use rabbit_ct_broker_helpers:stop_node/2
... instead of `rabbitmqctl stop` in
`simple_confirm_availability_on_leader_change` testcase.

The problem with the latter is that it doesn't wait for the broker to
actually stop. Therefore we end up with an error when we try to restart
it after because the previous instance is still running.

(cherry picked from commit 152219052fd47a85c8bb4ab6e706915a31f72364)