Optimise QQ memory use
Take fewer release cursor snapshots points as the message backlog grows.

Also introduces a compacted form of the internal message header map
where initially it is only an integer representing the size of the
message body. Later when additional keys need to be added it is expanded
into a full map. This avoid creating and holding many individial maps
with just a size element.

[#169064158]