Add run-in-minikube make target that resolves all deps
Only enabled on Darwin (a.k.a. macOS, a.k.a. OS X) because it's the only
platform that we have tested. Contributions for other platforms are
welcome.

We should uninstall VirtualBox, Minikube & kubectl and test that the
target works from a clean slate.

We are using VirtualBox as the VM driver for Minikube because
https://github.com/kubernetes/minikube/issues/5951

Pair @michaelklishin