unit_inbroker_non_parallel_SUITE: Skip log file test if dir writable
If the supposedly non-writable directory is writable (e.g. we are
running the testsuite on Windows as Administrator), we skip this test.

On Unix, we now use `/` as the non-writable directory because
`/var/empty` may not exist. For instance, this is the case on Debian
Stretch AWS EC2 image.

(cherry picked from commit 28a41d37061b9c01f2f8e295d2d963cb1a07f36b)