stop_command.ex: Read PID file before stopping the node
Now that the node removes its PID file on exit, we need to read it
before stopping the node.

Otherwise, if the PID file was already removed when
`OsPid.read_pid_from_file()` is called, it will wait for the PID file to
appear again in an infinite loop.

This was found when testing the RHEL 6 package on CentOS 6 in CI.

(cherry picked from commit e64aea80a72d667b9ff389f034e57e7ee1181231)