rabbit: Notify systemd after marking RabbitMQ as ready
Otherwise, there is a small time window where systemd considers the
service as running, but `rabbit:is_running()` would return false.

For instance, if "rabbitmqctl status" is called during that window, it
would return an error saying that RabbitMQ must be started on the node.
This must be the situation hit during package testing in CI.