rabbit_feature_flags: Use new `feature_flags` log category
This allows use to turn on/off and change the log level of the feature
flags subsystem specifically

While here, add two messages to log the time it took to query module
attributes (i.e. to find supported feature flags) and to regenerate the
registry (i.e. generate, compile and reload the rabbit_ff_registry
module).